Move all phases over.
